refactor: Controller 추상화
