+++ src/main/java/Application.java
@@ -1,11 +1,11 @@
-import controller.AttendanceController;
+import controller.MainController;
 import view.InputView;
 import view.OutputView;
 
 public class Application {
 
     public static void main(String[] args) {
-        AttendanceController attendanceController = new AttendanceController(new InputView(), new OutputView());
-        attendanceController.run();
+        MainController mainController = new MainController(new InputView(), new OutputView());
+        mainController.run();
     }
 }
